+++ README.md
@@ -1,8 +1,8 @@
-# Redgate SQL CI Jenkins Plugin
+# Redgate DLM Automation Jenkins Plugin
 
 ## Introduction
 
-The Redgate SQL CI Plugin is an open-source plugin for using [Redgate SQL CI](http://documentation.red-gate.com/display/SCI1/SQL+CI+documentation) from within Jenkins. Four tasks are available:
+The Redgate DLM Automation Plugin is an open-source plugin for using [Redgate DLM Automation](http://documentation.red-gate.com/display/SCI1/SQL+CI+documentation) from within Jenkins. Four tasks are available:
 
 1. Build your database from a [Redgate SQL Source Control](http://documentation.red-gate.com/display/SOC3/SQL+Source+Control+3+documentation) source folder to a nuget package.
 2. Test your database by running the tSQLt tests in the package.
@@ -14,7 +14,7 @@ If you just want to use the plugin, follow these instructions:
 
 1. Open your Jenkins.
 2. Go to Manage Jenkins > Manage Plugins> Available and search for Redgate.
-3. Tick the Redgate SQL CI Plugin, and click 'Install Without Restart'
+3. Tick the Redgate DLM Automation Plugin, and click 'Install Without Restart'
 
 ## How to build/debug the plugin.
 A basic tutorial for developing plugins is at [https://wiki.jenkins-ci.org/display/JENKINS/Plugin+tutorial](https://wiki.jenkins-ci.org/display/JENKINS/Plugin+tutorial).
@@ -26,7 +26,7 @@ Simply:
 3. Open a command prompt at the repository root directory. Then run:
     * `set MAVEN_OPTS=-Xdebug -Xrunjdwp:transport=dt_socket,server=y,address=8000,suspend=n`
     * ` mvn hpi:run`
-A development copy of Jenkins starts-up with the Redgate SQL CI plugin loaded.
+A development copy of Jenkins starts-up with the Redgate DLM Automation plugin loaded.
 
 [JetBrains IntelliJ IDEA](https://www.jetbrains.com/idea/) is a good environment for developing and debugging Jenkins plugins. There is a free community edition.
