+++ pom.xml
@@ -14,9 +14,9 @@
   <version>1.0.13-SNAPSHOT</version>
   <packaging>hpi</packaging>
 
-  <name>Redgate SQL CI</name>
+  <name>Redgate DLM Automation</name>
   <url>https://wiki.jenkins-ci.org/display/JENKINS/Redgate+SQL+CI+Plugin</url>
-  <description>Build, test, sync and publish your databases using Redgate SQL CI.</description>
+  <description>Build, test, sync and publish your databases using Redgate DLM Automation.</description>
   <licenses>
     <license>
       <name>Apache 2</name>
