+++ src/main/java/redgatedlmautomation/DlmDashboard.java
@@ -1,4 +1,4 @@
-package redgatesqlci;
+package redgatedlmautomation;
 
 import org.kohsuke.stapler.DataBoundConstructor;
