+++ src/main/java/redgatedlmautomation/GenerateTestData.java
@@ -1,4 +1,4 @@
-package redgatesqlci;
+package redgatedlmautomation;
 
 import org.kohsuke.stapler.DataBoundConstructor;
