+++ src/main/resources/redgatedlmautomation/BuildBuilder/global.jelly
