+++ src/main/resources/redgatedlmautomation/PublishBuilder/global.jelly
