+++ src/main/resources/redgatedlmautomation/PublishBuilder/help-packageVersion.html
