+++ src/main/resources/redgatedlmautomation/SyncBuilder/global.jelly
