+++ src/main/resources/redgatedlmautomation/TestBuilder/global.jelly
